+++ NEWS.rst
@@ -5,8 +5,9 @@ Lists of changes between versions.
 
 Latest
 ------
+* Minor: Added bouillon.git.working-directory-updated() which check that local is up to date with remote.
 * Minor: news_files, build_steps, lint_steps, test_steps can be set to nothing to disable them.
-* Minor. Unified cli arguments.
+* Minor: Unified cli arguments.
 
 2.4.0
 -----